+++ MathsVisualisationTool/MathsVisualisationTool/Properties/AssemblyInfo.cs
