+++ MathsVisualisationTool/MathsVisualisationTool/Properties/Resources.resx
