+++ MathsVisualisationTool/MathsVisualisationTool/interpreter/Lexer.cs
