+++ MathsVisualisationTool/MathsVisualisationTool/resources/App.xaml
